greatly simplify scheduler callbacks, handle all pending events in a tick
callbacks are stored as plain functions now (in the callbacks table)

when an event arrives, one thread is created and scheduled for each callback
registered for the corresponding evtype. this implementation removes the
limitation that only one event callback can be scheduled per tick.

the distinction between foreground and background threads is also gone - as
background threads were only used for callbacks, this concept is not needed
anymore.