integrate SDL event loop into the scheduler
upon loading the `sdl2` module, we replace sched.poller_factory with
an augmented version of itself. the only change is the implementation
of wait(timeout, process): instead of directly polling the scheduler's
fd, we use the SDL_WaitEvent() and SDL_WaitEventTimeout() functions to
wait for SDL events.

upon initializing the `sdl2` scheduler module, we start a C thread
which polls the scheduler's fd in an infinite loop. when the
scheduler's fd becomes readable, the poller thread sends a
user-defined SDL event (ZZ_SCHED_FD_POLLIN_EVENT) to the SDL event
loop (but does not read anything from the scheduler's fd yet).

when the sdl2 poller gets this SDL event, it calls the original
poller's wait() with a zero timeout (at this point we can be sure that
there are events to process so there is no need to wait).

the major advantage of this solution is that SDL events are delivered
even if the event loop blocks. with the previous solution - polling
SDL events in the sdl2 scheduler module's tick() method - we had to
keep the event loop running to get SDL events, which may be OK in a
game which needs to refresh itself 60 times per second anyway, but not
OK in an app which spends most of its time waiting for user input.

the major disadvantage is the requirement to link to the SDL2 lib at
compilation time, even if the sdl2 module is not used at all.