selective_redraw for ui containers
in some cases, we want to render the screen continuously, 60 times per second

in other cases, we want to render only when something changed

this feature supports the second case

by setting selective_redraw=true on a container, we ask it to draw
only those descendant widgets at the next draw() call which explicitly
asked for redraw by calling their Widget:redraw() method

Widget:redraw() bubbles up to the nearest container with
selective_redraw=true, that container pushes the widget onto its
redraw_list

windows may also have a redraw_trigger which can be fired when it's
time to redraw. the window render loop checks for the existence of a
redraw_trigger on the window. if the window has one, the sleep phase
polls the redraw_trigger so the next draw will happen only when the
redraw_trigger fires

if a container has a parent, it forwards redraw() calls to the parent,
otherwise it checks for a redraw_trigger on itself and fires it if
found (currently this is only useful on top-level windows)

typically, the top-level window will have selective_redraw and also a
redraw_trigger. when widgets down the hierarchy call redraw(), this
call will bubble up to the top-level window where the widget will be
placed onto the redraw_list and the top-level window's redraw_trigger
will be fired which makes the render loop redraw the window

several widget:redraw() calls within one cycle of the scheduler loop
result in several redraw_trigger fires but that's not a problem as
these are coalesced into one event