poll sched fd once per SDL poll cycle
writing to an eventfd which was registered in the scheduler's epoll
set caused the epoll_wait() in the SDL poller to return not just once,
but repeatedly, many times. each of these returns generated a new
SCHED_FD_POLLIN_EVENT in the SDL event queue, which piled up and choke
the Lua event processing loop.

first I tried to set the EPOLLET flag for sched_fd, in the hope that
then it would signal only the first change in the epoll_set.
unfortunately, this did not work.

finally I set the EPOLLONESHOT flag on sched_fd. this means that after
epoll_wait sees activity on sched_fd, it disables further event
notifications on it (so only the first event will be transformed into
a SCHED_FD_POLLIN_EVENT). to continue polling sched_fd, the main Lua
thread fires a new poll_trigger at the start of poller:wait() which is
handled in the poller thread by re-enabling the monitoring of sched_fd.